+++ swarm-derive/tests/test.rs
@@ -21,6 +21,7 @@
 use futures::prelude::*;
 use libp2p::swarm::{NetworkBehaviour, SwarmEvent};
 use libp2p_swarm_derive::*;
+use std::fmt::Debug;
 
 /// Small utility to check that a type implements `NetworkBehaviour`.
 #[allow(dead_code)]
@@ -82,8 +83,6 @@ fn three_fields() {
         ping: libp2p::ping::Ping,
         identify: libp2p::identify::Identify,
         kad: libp2p::kad::Kademlia<libp2p::kad::record::store::MemoryStore>,
-        #[behaviour(ignore)]
-        foo: String,
     }
 
     #[allow(dead_code)]
@@ -103,62 +102,7 @@ fn three_fields() {
 }
 
 #[test]
-fn three_fields_non_last_ignored() {
-    #[allow(dead_code)]
-    #[derive(NetworkBehaviour)]
-    struct Foo {
-        ping: libp2p::ping::Ping,
-        #[behaviour(ignore)]
-        identify: String,
-        kad: libp2p::kad::Kademlia<libp2p::kad::record::store::MemoryStore>,
-    }
-
-    #[allow(dead_code)]
-    #[allow(unreachable_code)]
-    fn foo() {
-        let _out_event: <Foo as NetworkBehaviour>::OutEvent = unimplemented!();
-        match _out_event {
-            FooEvent::Ping(libp2p::ping::Event { .. }) => {}
-            FooEvent::Kad(event) => {
-                let _: libp2p::kad::KademliaEvent = event;
-            }
-        }
-    }
-}
-
-#[test]
-fn custom_polling() {
-    #[allow(dead_code)]
-    #[derive(NetworkBehaviour)]
-    #[behaviour(poll_method = "foo")]
-    struct Foo {
-        ping: libp2p::ping::Ping,
-        identify: libp2p::identify::Identify,
-    }
-
-    impl Foo {
-        fn foo(
-            &mut self,
-            _: &mut std::task::Context,
-            _: &mut impl libp2p::swarm::PollParameters,
-        ) -> std::task::Poll<
-            libp2p::swarm::NetworkBehaviourAction<
-                <Self as NetworkBehaviour>::OutEvent,
-                <Self as NetworkBehaviour>::ConnectionHandler,
-            >,
-        > {
-            std::task::Poll::Pending
-        }
-    }
-
-    #[allow(dead_code)]
-    fn foo() {
-        require_net_behaviour::<Foo>();
-    }
-}
-
-#[test]
-fn custom_event_no_polling() {
+fn custom_event() {
     #[allow(dead_code)]
     #[derive(NetworkBehaviour)]
     #[behaviour(out_event = "MyEvent")]
@@ -181,54 +125,6 @@ fn custom_event_no_polling() {
     impl From<libp2p::identify::IdentifyEvent> for MyEvent {
         fn from(event: libp2p::identify::IdentifyEvent) -> Self {
             MyEvent::Identify(event)
-        }
-    }
-
-    #[allow(dead_code)]
-    fn foo() {
-        require_net_behaviour::<Foo>();
-    }
-}
-
-#[test]
-fn custom_event_and_polling() {
-    #[allow(dead_code)]
-    #[derive(NetworkBehaviour)]
-    #[behaviour(poll_method = "foo", out_event = "MyEvent")]
-    struct Foo {
-        ping: libp2p::ping::Ping,
-        identify: libp2p::identify::Identify,
-    }
-
-    enum MyEvent {
-        Ping(libp2p::ping::PingEvent),
-        Identify(libp2p::identify::IdentifyEvent),
-    }
-
-    impl From<libp2p::ping::PingEvent> for MyEvent {
-        fn from(event: libp2p::ping::PingEvent) -> Self {
-            MyEvent::Ping(event)
-        }
-    }
-
-    impl From<libp2p::identify::IdentifyEvent> for MyEvent {
-        fn from(event: libp2p::identify::IdentifyEvent) -> Self {
-            MyEvent::Identify(event)
-        }
-    }
-
-    impl Foo {
-        fn foo(
-            &mut self,
-            _: &mut std::task::Context,
-            _: &mut impl libp2p::swarm::PollParameters,
-        ) -> std::task::Poll<
-            libp2p::swarm::NetworkBehaviourAction<
-                <Self as NetworkBehaviour>::OutEvent,
-                <Self as NetworkBehaviour>::ConnectionHandler,
-            >,
-        > {
-            std::task::Poll::Pending
         }
     }
 
@@ -272,10 +168,27 @@ fn custom_event_mismatching_field_names() {
 }
 
 #[test]
+fn bound() {
+    #[allow(dead_code)]
+    #[derive(NetworkBehaviour)]
+    struct Foo<T: Copy + NetworkBehaviour>
+    where
+        <T as NetworkBehaviour>::OutEvent: Debug,
+    {
+        ping: libp2p::ping::Ping,
+        bar: T,
+    }
+}
+
+#[test]
 fn where_clause() {
     #[allow(dead_code)]
     #[derive(NetworkBehaviour)]
-    struct Foo<T: Copy + NetworkBehaviour> {
+    struct Foo<T>
+    where
+        T: Copy + NetworkBehaviour,
+        <T as NetworkBehaviour>::OutEvent: Debug,
+    {
         ping: libp2p::ping::Ping,
         bar: T,
     }
@@ -421,58 +334,19 @@ fn custom_event_with_either() {
 }
 
 #[test]
-fn mixed_field_order() {
-    struct Foo {}
-
-    #[derive(NetworkBehaviour)]
-    pub struct Behaviour {
-        #[behaviour(ignore)]
-        _foo: Foo,
-        _ping: libp2p::ping::Ping,
-        #[behaviour(ignore)]
-        _foo2: Foo,
-        _identify: libp2p::identify::Identify,
-        #[behaviour(ignore)]
-        _foo3: Foo,
-    }
-
-    #[allow(dead_code)]
-    fn behaviour() {
-        require_net_behaviour::<Behaviour>();
-    }
-}
-
-#[test]
-fn event_process() {
+fn generated_out_event_derive_debug() {
     #[allow(dead_code)]
     #[derive(NetworkBehaviour)]
-    #[behaviour(event_process = true)]
     struct Foo {
         ping: libp2p::ping::Ping,
-        identify: libp2p::identify::Identify,
     }
 
-    impl libp2p::swarm::NetworkBehaviourEventProcess<libp2p::identify::IdentifyEvent> for Foo {
-        fn inject_event(&mut self, _: libp2p::identify::IdentifyEvent) {}
+    fn require_debug<T>()
+    where
+        T: NetworkBehaviour,
+        <T as NetworkBehaviour>::OutEvent: Debug,
+    {
     }
 
-    impl libp2p::swarm::NetworkBehaviourEventProcess<libp2p::ping::PingEvent> for Foo {
-        fn inject_event(&mut self, _: libp2p::ping::PingEvent) {}
-    }
-
-    #[allow(dead_code, unreachable_code)]
-    fn bar() {
-        require_net_behaviour::<Foo>();
-
-        let mut _swarm: libp2p::Swarm<Foo> = unimplemented!();
-
-        let _ = async {
-            loop {
-                match _swarm.select_next_some().await {
-                    SwarmEvent::Behaviour(()) => break,
-                    _ => {}
-                }
-            }
-        };
-    }
+    require_debug::<Foo>();
 }