+++ misc/metrics/CHANGELOG.md
@@ -1,4 +1,18 @@
-# 0.8.0 [unreleased]
+# 0.9.0 [unreleased]
+
+- Update to `libp2p-swarm` `v0.39.0`.
+
+- Update to `libp2p-dcutr` `v0.6.0`.
+
+- Update to `libp2p-ping` `v0.39.0`.
+
+- Update to `libp2p-identify` `v0.39.0`.
+
+- Update to `libp2p-relay` `v0.12.0`.
+
+- Update to `libp2p-kad` `v0.40.0`.
+
+# 0.8.0
 
 - Update to `libp2p-swarm` `v0.38.0`.
 
@@ -16,9 +30,9 @@
 
 - Update to `libp2p-core` `v0.35.0`.
 
-- Update to `prometheus-client` `v0.17.0`. See [PR 2761].
+- Update to `prometheus-client` `v0.18.0`. See [PR 2822].
 
-[PR 2761]: https://github.com/libp2p/rust-libp2p/pull/2761/
+[PR 2822]: https://github.com/libp2p/rust-libp2p/pull/2761/
 
 [PR 2734]: https://github.com/libp2p/rust-libp2p/pull/2734/