+++ protocols/relay/CHANGELOG.md
@@ -1,4 +1,8 @@
-# 0.11.0 [unreleased]
+# 0.12.0 [unreleased]
+
+- Update to `libp2p-swarm` `v0.39.0`.
+
+# 0.11.0
 
 - Update to `libp2p-swarm` `v0.38.0`.